Remove timeout from gen_tcp:recv
This timeout value imposes a pointless upper bound on statement execution
time. The TCP layer should be allowed to wait forever on MySQL packets
and reset as needed by a monitoring process. That's what happens e.g.
in the "monitor work" pattern used in emysql.erl -- a statement that takes
too long is killed via a TCP connection reset.